+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/model_comparators/comparison/ElementComparator.java
@@ -1,15 +1,14 @@
-package com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison;
+package com.samares_engineering.omf.omf_core_framework.model_comparators.comparison;
 
-import com.nomagic.uml2.ext.magicdraw.mdprofiles.ExtensionEnd;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.diffmanagement.DiffManager;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.helpers.ComparatorUtils;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.helpers.ElementMatcher;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.propertycomparators.AttributeComparator;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.propertycomparators.ReferenceComparator;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.propertycomparators.TaggedValueComparator;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.diffdata.DiffKind;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.diffdata.dataclasses.ElementDiff;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.diffdata.dataclasses.PropertyDiff;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.diffmanagement.DiffManager;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.helpers.ComparatorUtils;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.helpers.ElementMatcher;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.propertycomparators.AttributeComparator;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.propertycomparators.ReferenceComparator;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.propertycomparators.TaggedValueComparator;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.diffdata.DiffKind;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.diffdata.dataclasses.ElementDiff;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.diffdata.dataclasses.PropertyDiff;
 import com.nomagic.magicdraw.tests.common.comparators.ModelComparatorFilter;
 import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Element;