+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/model_comparators/comparison/FilterManager.java
@@ -1,4 +1,4 @@
-package com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison;
+package com.samares_engineering.omf.omf_core_framework.model_comparators.comparison;
 
 import com.nomagic.magicdraw.tests.common.comparators.ModelComparatorFilter;
 import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Element;