+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/model_comparators/comparison/LegacyLogger.java
@@ -1,6 +1,6 @@
-package com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison;
+package com.samares_engineering.omf.omf_core_framework.model_comparators.comparison;
 
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.helpers.ComparatorUtils;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.helpers.ComparatorUtils;
 import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Element;
 import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.TaggedValue;