+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/model_comparators/comparison/diffmanagement/DiffManager.java
@@ -1,8 +1,8 @@
-package com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.diffmanagement;
+package com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.diffmanagement;
 
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.comparison.ElementComparator;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.diffdata.DiffKind;
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.diffdata.dataclasses.ElementDiff;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.comparison.ElementComparator;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.diffdata.DiffKind;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.diffdata.dataclasses.ElementDiff;
 import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Element;
 
 import javax.annotation.Nonnull;