+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/model_comparators/diffdata/dataclasses/ElementDiff.java
@@ -1,6 +1,6 @@
-package com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.diffdata.dataclasses;
+package com.samares_engineering.omf.omf_core_framework.model_comparators.diffdata.dataclasses;
 
-import com.samares_engineering.omf.omf_test_framework.projectcomparator.model_comparators.diffdata.DiffKind;
+import com.samares_engineering.omf.omf_core_framework.model_comparators.diffdata.DiffKind;
 import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Element;
 
 import javax.annotation.Nullable;