+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+### Changed
+- OMF Comparator was moved from omf-test-framework to omf-core-framework, and dependencies from omf-test were removed
+
 
 ## [1.2.0] - 2024-01-17
 ### Added